move COLOR into its own header; move uibox_* functions and ui related globals into UI_CONTROL.h; remove SDL_FontCache; rename UPDATE_INPUT to SYSTEM_INPUT_UPDATE; rename SYSTEM_UIBOX_CONTROL to SYSTEM_UIBOX_UPDATE
