+++ src/SYSTEM.h
@@ -8,7 +8,10 @@ void INIT_SDL();
 SDL_Window* INIT_WINDOW();
 SDL_Renderer* INIT_RENDERER(SDL_Window* WINDOW);
 SDL_Texture* INIT_FONT(SDL_Renderer* renderer);
-void UPDATE_INPUT();
+
+void SYSTEM_SHUTDOWN(SDL_Window* WINDOW);
+
+void SYSTEM_INPUT_UPDATE();
 
 void SYSTEM_BRUSH_UPDATE();
 void SYSTEM_LAYER_UPDATE();
